requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pydantic